Address review: extract na-cli selection, clarify comments, log exec failures
- Move the na-cli binary selection and eBPF collection code out of
  networking.go into networking_cli.go, matching the existing
  one-concept-per-file layout and the networking_cli_internal_test.go
  name.
- Fix the CLI search-order comment: /opt/cni/bin is the steady-state CNI
  daemonset location on EC2 AL2023, not a legacy path. Cite the
  single-installed-binary family-confirmation reasoning on the default
  and v6-named branches.
- Write loaded-ebpfdata execution failures into the bundle so the
  "*** EBPF loaded data ***" header is never left dangling and a denied
  exec on Bottlerocket is legible in the artifact rather than only in the
  returned error.
