+++ pkg/log_collector/collect/networking.go
@@ -8,7 +8,6 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 
 	"k8s.io/client-go/tools/clientcmd"
@@ -192,155 +191,4 @@ func apiServerConnectivity(acc *Accessor) error {
 		}
 	}
 	return merr
-}
-
-// na-cli binary names. The v4 and v6 builds decode only their own family's map
-// keys. EC2 installs the family-correct build under the default name; Auto Mode
-// bakes both (default = v4), so a published symlink disambiguates.
-const (
-	naCLIDefault = "aws-eks-na-cli"
-	naCLIv6Named = "aws-eks-na-cli-v6"
-)
-
-// networkPolicyCLIDirs are searched in order. On Bottlerocket the /opt/cni/bin
-// copy is SELinux-labeled cni_exec_t, which the agent's domain (system_t) cannot
-// execute; the /usr/bin copy is os_t, which it can. So /usr/bin is tried first,
-// with /opt/cni/bin as a fallback for images that predate the /usr/bin install.
-var networkPolicyCLIDirs = []string{"/usr/bin", "/opt/cni/bin"}
-
-// naCLIPublishedDir is where the network policy agent publishes a symlink to the
-// family-correct build at startup (it knows the family from its own config). In
-// /run because /usr is immutable; absent until the agent starts, cleared on
-// reboot. Absence means the family is unconfirmed and map decoding is skipped.
-const naCLIPublishedDir = "/run/aws-network-policy-agent"
-
-// findNetworkPolicyCLI returns the Root-prefixed path of the named CLI if it
-// exists in one of networkPolicyCLIDirs.
-func findNetworkPolicyCLI(acc *Accessor, name string) (string, bool) {
-	for _, dir := range networkPolicyCLIDirs {
-		p := filepath.Join(acc.cfg.Root, dir, name)
-		if _, err := os.Stat(p); err == nil {
-			return p, true
-		}
-	}
-	return "", false
-}
-
-// findPublishedNetworkPolicyCLI returns the published symlink if it exists and
-// its target exists, both resolved under acc.cfg.Root. The link target is an
-// absolute path, so it is re-rooted under Root before the existence check rather
-// than followed by os.Stat (which would resolve against the wrong namespace when
-// Root != "/"). A dangling link reads as absent, so decoding is skipped.
-func findPublishedNetworkPolicyCLI(acc *Accessor) (string, bool) {
-	link := filepath.Join(acc.cfg.Root, naCLIPublishedDir, naCLIDefault)
-	target, err := os.Readlink(link)
-	if err != nil {
-		return "", false
-	}
-	// Re-root an absolute target under Root; a relative target is already
-	// relative to the link's directory.
-	resolved := target
-	if filepath.IsAbs(target) {
-		resolved = filepath.Join(acc.cfg.Root, target)
-	} else {
-		resolved = filepath.Join(filepath.Dir(link), target)
-	}
-	if _, err := os.Stat(resolved); err != nil {
-		return "", false
-	}
-	return link, true
-}
-
-// selectNetworkPolicyCLI chooses which na-cli binary may decode eBPF maps.
-// Returns ok=false to skip decoding when no CLI is installed or the family is
-// unconfirmed. loaded-ebpfdata (family-agnostic) is collected regardless.
-func selectNetworkPolicyCLI(acc *Accessor) (cliPath string, reason string, ok bool) {
-	publishedCLI, havePublished := findPublishedNetworkPolicyCLI(acc)
-	defaultCLI, haveDefault := findNetworkPolicyCLI(acc, naCLIDefault)
-	v6NamedCLI, haveV6Named := findNetworkPolicyCLI(acc, naCLIv6Named)
-	return chooseNetworkPolicyCLI(publishedCLI, havePublished, defaultCLI, haveDefault, v6NamedCLI, haveV6Named)
-}
-
-// chooseNetworkPolicyCLI is the pure selection decision, split from the
-// filesystem side for unit testing. The published symlink is the only family
-// signal; there is no family query and no decode probe.
-func chooseNetworkPolicyCLI(publishedCLI string, havePublished bool, defaultCLI string, haveDefault bool, v6NamedCLI string, haveV6Named bool) (cliPath string, reason string, ok bool) {
-	switch {
-	case havePublished:
-		return publishedCLI, "NPA-published family-correct symlink present: using it", true
-	case !haveDefault && !haveV6Named:
-		return "", fmt.Sprintf("neither %s nor %s found in %v; network policy agent not installed", naCLIDefault, naCLIv6Named, networkPolicyCLIDirs), false
-	case haveDefault && !haveV6Named:
-		return defaultCLI, "single binary (aws-eks-na-cli) present — installed family-correct by the CNI daemonset: using it", true
-	case !haveDefault && haveV6Named:
-		return v6NamedCLI, fmt.Sprintf("only %s present (unexpected layout); using it", naCLIv6Named), true
-	}
-	return "", "both na-cli builds present but no NPA-published symlink; family unconfirmed, skipping map dumps", false
-}
-
-// networkPolicyEbpfInfo dumps the network-policy eBPF loaded programs and map
-// contents via aws-eks-na-cli. loaded-ebpfdata is family-agnostic and runs with
-// whichever binary is installed; the per-map dump binary is chosen by
-// selectNetworkPolicyCLI.
-func networkPolicyEbpfInfo(acc *Accessor) error {
-	ebpfDataFile := "networking/ebpf-data.txt"
-
-	listCLI, listOK := findNetworkPolicyCLI(acc, naCLIDefault)
-	if !listOK {
-		listCLI, listOK = findNetworkPolicyCLI(acc, naCLIv6Named)
-	}
-	if !listOK {
-		return acc.appendOutput(ebpfDataFile, []byte(fmt.Sprintf("*** %s/%s not present, skipping eBPF map collection ***\n", naCLIDefault, naCLIv6Named)))
-	}
-
-	if err := acc.appendOutput(ebpfDataFile, []byte("*** EBPF loaded data ***\n")); err != nil {
-		return fmt.Errorf("failed to append output to %s: %w", ebpfDataFile, err)
-	}
-	loaded, err := acc.Command(listCLI, "ebpf", "loaded-ebpfdata").CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("failed to execute %s: %w", listCLI, err)
-	}
-	if err := acc.appendOutput(ebpfDataFile, loaded); err != nil {
-		return fmt.Errorf("failed to append output to %s: %w", ebpfDataFile, err)
-	}
-
-	// Record the selection reason even on skip: it is the bundle's diagnostic.
-	cliPath, reason, ok := selectNetworkPolicyCLI(acc)
-	if err := acc.appendOutput(ebpfDataFile, []byte(fmt.Sprintf("*** network-policy CLI selection: %s ***\n", reason))); err != nil {
-		return fmt.Errorf("failed to append CLI selection to %s: %w", ebpfDataFile, err)
-	}
-	if !ok {
-		return nil
-	}
-
-	ebpfMapFile := "networking/ebpf-maps-data.txt"
-	mapIDs := uniqueMapIDs(string(loaded))
-	var merr error
-	for _, mapID := range mapIDs {
-		merr = errors.Join(merr,
-			acc.appendOutput(ebpfMapFile, []byte(fmt.Sprintf("*** EBPF map data for Map ID: %s ***\n", mapID))),
-			acc.CommandOutput([]string{cliPath, "ebpf", "dump-maps", mapID}, ebpfMapFile, CommandOptionsAppend|CommandOptionsNoStderr),
-		)
-	}
-	return merr
-}
-
-// uniqueMapIDs extracts the distinct "Map ID: <n>" values from loaded-ebpfdata
-// output, sorted, so map collection is deterministic run-to-run.
-func uniqueMapIDs(loaded string) []string {
-	seen := make(map[string]bool)
-	var ids []string
-	for _, line := range strings.Split(loaded, "\n") {
-		idx := strings.Index(line, "Map ID:")
-		if idx < 0 {
-			continue
-		}
-		id := strings.TrimSpace(line[idx+len("Map ID:"):])
-		if id != "" && !seen[id] {
-			seen[id] = true
-			ids = append(ids, id)
-		}
-	}
-	sort.Strings(ids)
-	return ids
 }
